+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v1.1.0 (in development)
 -----------------------
 - `"buildver"` is now `None`/`null` instead of the empty string when there is
   no build tag in the wheel's filename
+- Validation of `RECORD` files is now done by `wheel-inspect` instead of
+  `distlib` for more descriptive error messages
 
 v1.0.0 (2018-10-12)
 -------------------